Cleaning up and reorganizing unit tests
